separating functions, extracting code into separate files
