+++ api/src/models/user.ts
@@ -4,4 +4,6 @@ const userSchema: mongoose.Schema = new mongoose.Schema({
     name: String,
 });
 
-export default mongoose.model('User', userSchema);
+const User = mongoose.model('User', userSchema);
+
+export default User;
